Add files via upload
EduConsultPro Institute is a leading educational institution offering a variety of courses and programs in diverse fields. With a growing number of prospective students seeking admission each year, the institute faces challenges in managing the admission process, students enquiry, and expert consulting services efficiently. To address these challenges, EduConsultPro Institute decides to leverage Salesforce CRM to streamline the admission process and enhance the overall experience for both students and admissions staff.

The use case focuses on the admission process for prospective students interested in enrolling in courses and programs offered by EduConsultPro Institute. The goal is to provide a seamless and transparent experience for students while enabling admissions staff to efficiently review and process admission applications, students enquiry and case management.